Add variations for cliches
If a cliche can be produced using mutiple individual samples,
multiple samples will be compiled. In additon to that a pattern
can be flagged "random" (which is default for cliches), so that
repeat_index is ignored and a random sample is taken instead.

this will add much more variations to the output file.